+++ test/unit-tests/json/replacer.test.js
@@ -72,6 +72,20 @@ describe('replacer', function () {
     assert.deepStrictEqual(JSON.stringify(u, replacer), json)
   })
 
+  it('should stringify a Unit with a value only', function () {
+    const u = new math.Unit(5)
+    const json = '{"mathjs":"Unit","value":5,"unit":null,"fixPrefix":false}'
+    assert.deepStrictEqual(JSON.stringify(u), json)
+    assert.deepStrictEqual(JSON.stringify(u, replacer), json)
+  })
+
+  it('should stringify a Unit without a value', function () {
+    const u = new math.Unit(null, 'cm')
+    const json = '{"mathjs":"Unit","value":null,"unit":"cm","fixPrefix":false}'
+    assert.deepStrictEqual(JSON.stringify(u), json)
+    assert.deepStrictEqual(JSON.stringify(u, replacer), json)
+  })
+
   it('should stringify a Matrix, dense', function () {
     const m = math.matrix([[1, 2], [3, 4]], 'dense')
     const json = '{"mathjs":"DenseMatrix","data":[[1,2],[3,4]],"size":[2,2]}'