+++ HISTORY.md
@@ -1,5 +1,10 @@
 # History
 
+# unpublished changes since 12.4.2
+
+- Fix: serialization of Units without a value, see #1240.
+
+
 # 2024-04-24, 12.4.2
 
 - Fix #3192: function `isNaN` returns `false` for `NaN` units in a matrix or